fix: Resolve Tailwind Configuration Mismatch
Fixed a bug in the Flipdish app where the Tailwind CSS build was not configured correctly. Resolved mismatches in the settings to ensure a consistent and accurate Tailwind configuration.
