+++ src/App.tsx
@@ -1,16 +1,12 @@
-import "./App.css";
-import Header from "../@/Header";
-import {apiDataFetch} from "../@/lib/apiDataFetch"
+import './App.css'
+import Header from './components/Header'
 
 function App() {
-  const { menuData, loading, error } = apiDataFetch();
-  console.log("App Component Log:", menuData, loading, error);
-
   return (
     <>
-      <Header />
+     <Header /> 
     </>
-  );
+  )
 }
 
-export default App;
+export default App
