+++ vite.config.ts
@@ -1,9 +1,10 @@
 import path from "path"
-import ReactRefresh from "@vitejs/plugin-react";
-import { defineConfig } from "vite"
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
 
+// https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [ReactRefresh()],
+  plugins: [react()],
   resolve: {
     alias: {
       "@": path.resolve(__dirname, "./src"),
